[code] termination without kick timer
This function removes the KICK timer and solves most of the
issues that the timer removal showed.

First, SHUTDOWN was not sent on client sockets, which could make
some clients lose last netorcai messages when netorcai finished.
Now, CloseWrite() should be called before any Socket Close(),
which should solve this issue.

Second, the Cleanup function directly sent Kick messages on Client
sockets, which caused a race condition with the goroutine in charge
of the client socket (in handlePlayerOrVisu).
Now, Cleanup tells the goroutine to do the KICK, avoiding this race
condition.

Third, some goroutines were still running at netorcai's termination.
This was not so important before because Cleanup directly kicked
clients, but as this is no longer the case, clean termination is now
needed.
This is solved by the end of all goroutines managing clients in
netorcai's main program.
This new behavior can cause a deadlock on netorcai's termination,
but as tests are quite thorough I hope no deadlock went through.

Fourth, some code removed clients from the global data structure
in an incorrect way, leading to a deadlock on Cleanup.